refactor: setlist관련 쿼리 구조 변경
